+++ plugins/quetz_tos/quetz_tos/db_models.py
@@ -6,12 +6,21 @@ from sqlalchemy.orm import backref, relationship
 from quetz.db_models import UUID, Base
 
 
+class TermsOfServiceFile(Base):
+    __tablename__ = 'quetz_tos_file'
+
+    id = Column(UUID, primary_key=True, default=lambda: uuid.uuid4().bytes)
+    filename = Column(String)
+    language = Column(String)
+    tos_id = Column(UUID, ForeignKey("quetz_tos.id"), primary_key=True)
+
+
 class TermsOfService(Base):
     __tablename__ = 'quetz_tos'
 
     id = Column(UUID, primary_key=True, default=lambda: uuid.uuid4().bytes)
     uploader_id = Column(UUID)
-    filename = Column(String)
+    files = relationship("TermsOfServiceFile")
     time_created = Column(DateTime, nullable=False, server_default=func.now())