+++ quetz/testing/utils.py
@@ -1,4 +1,6 @@
 import json
+import signal
+from multiprocessing import active_children
 
 from starlette.requests import Request as ASGIRequest
 from starlette.responses import Response as ASGIResponse
@@ -30,3 +32,18 @@ class AsyncPathMapDispatch:
             headers=headers,
         )
         await response(scope, receive, send)
+
+
+class Interrupt:
+    # Interrupt child when SIGALRM is received.
+    # Useful to kill the server when it is correctly launched, using a timeout.
+    def _handle_interrupt(self, signum, frame):
+        for p in active_children():
+            p.terminate()
+            p.join()
+
+    def __enter__(self):
+        signal.signal(signal.SIGALRM, self._handle_interrupt)
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        pass